+++ source/TestCommon/source/FunctionApp.TestCommon.Tests/Fixtures/ServiceBusListenerMockFixture.cs
@@ -18,7 +18,7 @@ using Energinet.DataHub.Core.TestCommon.Diagnostics;
 namespace Energinet.DataHub.Core.FunctionApp.TestCommon.Tests.Fixtures;
 
 /// <summary>
-/// This fixtures ensures we reuse <see cref="ConnectionString"/>
+/// This fixtures ensures we reuse <see cref="FullyQualifiedNamespace"/>
 /// so we only have to retrieve an access token and values in Key Vault one time.
 ///
 /// When testing the <see cref="ServiceBusListenerMock"/> we must create new queues/topics
@@ -30,13 +30,10 @@ public class ServiceBusListenerMockFixture
     public ServiceBusListenerMockFixture()
     {
         TestLogger = new TestDiagnosticsLogger();
-
-#pragma warning disable CS0618 // Type or member is obsolete
-        ConnectionString = SingletonIntegrationTestConfiguration.Instance.ServiceBusConnectionString;
-#pragma warning restore CS0618 // Type or member is obsolete
+        FullyQualifiedNamespace = SingletonIntegrationTestConfiguration.Instance.ServiceBusFullyQualifiedNamespace;
     }
 
     public ITestDiagnosticsLogger TestLogger { get; }
 
-    public string ConnectionString { get; }
+    public string FullyQualifiedNamespace { get; }
 }